Added Rooms to the Seeder. Renamed the room and hotel packages and moved the seeder package out of the hotel package.
