+++ backend/src/main/java/nl/itvitae/BookingApp/hotel/HotelRepository.java
@@ -1,7 +1,6 @@
-package nl.itvitae.BookingApp.hotels;
+package nl.itvitae.BookingApp.hotel;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
 @Repository
